// app-five/src/types/formActions.ts
export enum FormActions {
  SET_FORM = 'SET_FORM',
  SET_LOADING = 'SET_LOADING',
  SET_ERROR = 'SET_ERROR',
  SET_SELECTED_QUESTION = 'SET_SELECTED_QUESTION',
  ADD_QUESTION = 'ADD_QUESTION',
  UPDATE_QUESTION = 'UPDATE_QUESTION',
  DELETE_QUESTION = 'DELETE_QUESTION',
  REORDER_QUESTIONS = 'REORDER_QUESTIONS',
  UPDATE_FORM_META = 'UPDATE_FORM_META',
  RESET_FORM = 'RESET_FORM',
  SET_DIRTY = 'SET_DIRTY'
}